+++ src/manager/routes/general/email/templates/download-app-link.js
@@ -14,10 +14,10 @@ module.exports = function (payload, config) {
         email: payload.email,
         name: payload.name,
       },
+      sender: 'marketing',
       categories: ['download'],
       subject: `Free ${config.brand.name} download link for ${payload.name || 'you'}!`,
       template: 'main/misc/app-download-link',
-      group: 'marketing',
       copy: false,
       data: {},
     }